requirements.txt
python-dotenv==1.0.0
pymongo==4.6.1
telethon==1.32.1
motor==3.3.2
pyyaml==6.0.1
dnspython==2.4.2
pytz==2023.3
loguru==0.7.2
apscheduler==3.10.4